Allow to reject connections before the TLS handshake
Motivation:

The ConnectionAcceptor at the moment only allows to reject
connections after the TLS handshake is complete. While this provides
the user with additional information inside the TLS session, the
actual handshake process takes CPU resources and time which in
some cases might not be necessary or even undesired.

Modifications:

This changeset adds additional functionality to allow to reject
a connection even before the TLS handshake computes / the netty
pipeline does its work.

It does this by adding a new method to the ConnectionAcceptor
which takes reduced connection info as an argument (right now
only the remote socket info) and allows the implementation to
reject a connection early. It has a default passthrough
implementation for backwards compatibility reasons.

Result:

Added functionality which allows to reject incoming connections
quikly and potentially wasting less resources in the process.